refactor(front): refactors post page and adds components
