+++ front/src/lib/components/posts/Post.svelte
@@ -1,0 +1,112 @@
+<script>
+  import ListIcon from "~icons/gg/list";
+  import Code from "@components/content/Code.svelte";
+  import Footnotes from "@components/content/renderers/Footnotes.svelte";
+  import Link from "@components/content/renderers/Link.svelte";
+  import PostNav from "@components/posts/PostNav.svelte";
+  import PostHeading from "@components/content/renderers/PostHeading.svelte";
+  import slugger from "slugger";
+  import SvelteMarkdown from "svelte-markdown";
+
+  export let title;
+  export let content;
+  export let footnotes;
+  export let categories;
+  export let related;
+  export let pathname;
+
+  let toc;
+  $: if (content) toc = [];
+  function filterTokens(event) {
+    const tokens = event.detail.tokens;
+    const headings = tokens.filter((t) => t.type === "heading");
+    toc = [
+      ...toc,
+      ...headings.map((h) => ({
+        text: h.text,
+        link: `#${slugger(h.text)}`,
+      })),
+    ];
+  }
+
+  function handleParsed(event) {
+    filterTokens(event);
+  }
+
+  $: contents = toc ? [...toc] : false;
+
+  $: showAside = true;
+  $: asideLabel = showAside ? "Hide" : "Show";
+  function asideToggle() {
+    showAside = !showAside;
+  }
+
+  function togglePostNav() {
+    console.log("postnav clickseds");
+    const postNavCheckbox = document.getElementById("post-navigation");
+    postNavCheckbox.checked = postNavCheckbox.checked ? false : true;
+  }
+</script>
+
+<nav class="post-navigation">
+  <input type="checkbox" name="post-navigation" id="post-navigation" tabindex="-1" />
+  <label for="post-navigation" class="post-navigation-toggle"
+    ><ListIcon /></label
+  >
+  <div
+    class="post-navigation-list"
+    on:click={togglePostNav}
+    on:keydown={togglePostNav}
+  >
+    <PostNav {contents} {categories} {related} {pathname} />
+    <label for="post-navigation" class="post-navigation-toggle bottom"
+      ><ListIcon /></label
+    >
+  </div>
+</nav>
+<h1 class="post-title">{title}</h1>
+<article class="post-article" class:full={!showAside}>
+  {#if content}
+    {#each content as c}
+      {#if c.__component === "posts.text"}
+        <SvelteMarkdown
+          source={c.text}
+          renderers={{ link: Link, heading: PostHeading }}
+          on:parsed={handleParsed}
+        />
+      {/if}
+      {#if c.__component === "posts.code"}
+        {@const lines =
+          c?.highlightedLines?.split(",").map((item) => Number(item - 1)) ||
+          false}
+        <Code
+          text={c.code}
+          lang={c.language}
+          title={c?.title}
+          highlightedLines={lines
+            ? lines.sort((a, b) => {
+                return a - b;
+              })
+            : false}
+        />
+      {/if}
+      {#if c.__component === "posts.html"}
+        <div class="inline-html">
+          {@html c.html}
+        </div>
+      {/if}
+    {/each}
+  {/if}
+  {#if footnotes}
+    <footer class="footnotes-footer">
+      <Footnotes {footnotes} />
+    </footer>
+  {/if}
+</article>
+
+<!-- <button on:click={asideToggle} class="aside-toggle"
+  >{asideLabel} Sidebar</button
+> -->
+<aside class="post-aside" class:show={showAside}>
+  <PostNav {contents} {categories} {related} {pathname} />
+</aside>
