+++ front/src/lib/components/posts/PostsList.svelte
@@ -1,0 +1,5 @@
+<script>
+    
+</script>
+
+List of Posts
